requirements.txt
boto3==1.24.29
botocore==1.27.29
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask-Cors==3.0.10
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
